codec2: CB1 pts calulate error [2/2]
PD#SWPL-98454

Problem:
1 video will be fastfoward when play the
specific h265 interlace stream.pts calute
error.
2 CtsMediaV2TestCases
android.mediav2.cts.DecodeGlAccuracyTest
[18(c2.amlogic.hevc.decoder_video/hevc_2_4_3_true)]
3 android.mediav2.cts.AdaptivePlaybackTest error.

Solution:
1. change out time to uint64.
2. add C2StreamMaxBufferSizeInfo::input init value,
fixed CtsMediaV2TestCases error case. if not init this
value, csd input buf cannot malloc,because the init value of
inputbuf is a smaller value.
3. we need check "mResolutionChanging" value is true
or not.we ignore outbuf when "mResolutionChanging" is
true,this will fixed android.mediav2.cts.AdaptivePlaybackTest
error.

Verify:
ohm

+++ vendorcomponents/videodecoder/utils/C2VdecTunnelHelper.cpp
@@ -191,7 +191,7 @@ void C2VdecComponent::TunnelHelper::onFillVideoFrameTunnel2(int dmafd, bool rend
                     [id = info->mBlockId](const OutputBufferInfo& o) { return o.mBlockId == id;});
             if (pendingBuffer != mComp->mPendingBuffersToWork.end()) {
                 struct VideoTunnelRendererWraper::renderTime rendertime = {
-                    .mediaUs = pendingBuffer->mMediaTimeUs,
+                    .mediaUs = (int64_t)pendingBuffer->mMediaTimeUs,
                     .renderUs = systemTime(SYSTEM_TIME_MONOTONIC) / 1000,
                 };
                 sendOutputBufferToWorkTunnel(&rendertime);
